+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+### v2.5.x
+
+* Add restore
+* (Breaking <2.5.0) Change lil5/questions source
+
+### v2.4.x
+
+* Add documentation at readme
+* Fix bugs
+* Add after and before custom scripts
+
+### v2.3.x
+
+* Add async
+
+### v2.2.x
+
+* Add Installation in documentation
+* Fix bugs
+* Change legal license
+* Add glob
+* (Breaking <2.2.0) Merge feature reorder
+
+### v2.1.x
+
+* (Breaking <2.1.0) Changed standard to .rsynced.hjson
+* Add environment support
+
+
 ### v2.0.1
 
 * Add `init` (like npm init)
@@ -16,7 +45,7 @@
 * Remove `group` option.
 * Add `chmod` option.
 
-### v1.2.0 (Deprecated)
+### v1.2.0
 
 * Add `owner` option.
 * Add `group` option.